future/settings: Migrate sysinfo to use the new future/settings API
As part of the effort to remove default values scattered around the
code, this change only migrates the sysinfo plugin options to use the
new module. This will associate the --sysinfo command line option with
the 'enabled' configuration of the 'sysinfo.collect' section in the
configuration file. Also will associate the positional argument
'sysinfodir' to a key with the this name under the same section
mentioned above.

Signed-off-by: Beraldo Leal <bleal@redhat.com>