future/settings: Migrate config to use the new future/settings API
As part of the effort to remove default values scattered around the
code, this change only migrates the config plugin options to use the new
module. This will associate the --datadir command line option with the
'datadir' configuration of the 'config' section in the configuration
file. Also will associate --paginator option to a key with the this name
under the same section mentioned above.

Signed-off-by: Beraldo Leal <bleal@redhat.com>