future/settings: Migrate nrun to use the new future/settings API
As part of the effort to remove default values scattered around the
code, this change only migrates the nrun plugin options to use the new
module. This will associate the references, --disable-task-randomization
and --status-server to nrun.references, nrun.disable_task_randomization
and nrun.status_server.listen respectively on configuration file.

Signed-off-by: Beraldo Leal <bleal@redhat.com>